Move ReferenceCursor files into model/
