+++ src/actions/set_reference_cursor_action.cpp
@@ -4,7 +4,7 @@
 #include <fullscore/actions/set_reference_cursor_action.h>
 
 #include <fullscore/models/measure_grid.h>
-#include <fullscore/reference_cursor.h>
+#include <fullscore/models/reference_cursor.h>
