+++ src/components/measure_grid_render_component.cpp
@@ -12,8 +12,8 @@
 #include <fullscore/helpers/measure_grid_helper.h>
 #include <fullscore/models/note.h>
 #include <fullscore/models/measure_grid.h>
+#include <fullscore/models/reference_cursor.h>
 #include <fullscore/services/music_engraver.h>
-#include <fullscore/reference_cursor.h>
