+++ src/models/reference_cursor.cpp
@@ -1,7 +1,7 @@
 
 
 
-#include <fullscore/reference_cursor.h>
+#include <fullscore/models/reference_cursor.h>
 
 #include <fullscore/models/measure_grid.h>
